Move tests to `ParameterBeam` test file
